correction backend

+++ backend/controllers/comment.js
@@ -1,32 +1,54 @@
 const { Post, User, Comment } = require('../models/Index');
 
 exports.createComment = (req, res, next) => {
-    Comment.create({
-        content: req.body.content
+    Post.findByPk(req.params.id)
+    .then(post => {
+        Comment.create({
+            content: req.body.content,
+            UserId: req.token.userId,
+            postId: req.params.id
+        })
+        .then(() => res.status(201).json({ message: 'Commentaire publié !'}))
+        .catch(error => res.status(400).json({ error }));
     })
-    .then(() => res.status(201).json({ message: 'Commentaire publié !'}))
-    .catch(error => res.status(400).json({ error }));
+    .catch(error => res.status(404).json({ error }));
 }
 
 exports.modifyComment = (req, res, next) => {
-    Comment.update({
-        content: req.body.content
-    },
-    {
-        where: {
-            id: req.params.id
+    Comment.findByPk(req.params.id)
+    .then(comment => {
+        if (req.token.userId == comment.UserId || User.admin == true) {
+            Comment.update({
+                content: req.body.content
+            },
+            {
+                where: {
+                    id: req.params.id
+                }
+            })
+            .then(() => res.status(200).json({ message: 'Commentaire modifié !'}))
+            .catch(error => res.status(400).json({ error }));
+        } else {
+            res.status(403).json({ message: "Vous n'avez pas l'autorisation pour modifié ce commentaire !" })
         }
     })
-    .then(() => res.status(200).json({ message: 'Commentaire modifié !'}))
-    .catch(error => res.status(400).json({ error }));
+    .catch(error => res.status(404).json({ error }));
 }
 
 exports.deleteComment = (req, res, next) => {
-    Comment.destroy({
-        where: {
-            id: req.params.id
+    Comment.findByPk(req.params.id)
+    .then(comment => {
+        if (req.token.userId == comment.UserId) {
+            Comment.destroy({
+                where: {
+                    id: req.params.id
+                }
+            })
+            .then(() => res.status(200).json({ message: 'Commentaire supprimé !'}))
+            .catch(error => res.status(400).json({ error }));
+        } else {
+            res.status(403).json({ message: "Vous n'avez pas l'autorisation pour supprimer ce commentaire !" })
         }
     })
-    .then(() => res.status(200).json({ message: 'Commentaire supprimé !'}))
-    .catch(error => res.status(400).json({ error }));
+    .catch(error => res.status(404).json({ error }));
 }
